+++ packages/react-article-components/src/components/donation-box.js
@@ -1,15 +1,17 @@
 import React, { PureComponent } from 'react'
 import styled from 'styled-components'
+import DOMPurify from 'isomorphic-dompurify'
 // @twreporter
 import externalLinks from '@twreporter/core/lib/constants/external-links'
 import mq from '@twreporter/core/lib/utils/media-query'
 import { DONATION_LINK_ANCHOR } from '@twreporter/core/lib/constants/donation-link-anchor'
 import { colorGrayscale } from '@twreporter/core/lib/constants/color'
+import { TEN_YEAR_ANNIVERSARY } from '@twreporter/core/lib/constants/feature-flag'
 // constants
 import predefinedCss from '../constants/css'
 import typography from '../constants/typography'
 
-const _content = {
+const content = {
   title: '用行動支持報導者',
   desc: [
     '獨立的精神，是自由思想的條件。獨立的媒體，才能守護公共領域，讓自由的討論和真相浮現。',
@@ -19,13 +21,25 @@ const _content = {
   bt: '立即支持',
 }
 
+const newContent = {
+  title: '深度求真 眾聲同行',
+  desc: [
+    '獨立的精神，是自由思想的條件。獨立的媒體，才能守護公共領域，讓自由的討論和真相浮現。',
+    '在艱困的媒體環境，《報導者》堅持以非營利組織的模式投入公共領域的調查與深度報導。我們透過讀者的贊助支持來營運，不仰賴商業廣告置入，在獨立自主的前提下，穿梭在各項重要公共議題中。',
+    `今年是<a href="${externalLinks.tenYearAnniversary}" target="_blank" rel="noopener noreferrer">《報導者》成立十週年</a>，請支持我們持續追蹤國內外新聞事件的真相，度過下一個十年的挑戰。`,
+  ],
+  bt: '贊助支持',
+}
+
+const _content = TEN_YEAR_ANNIVERSARY ? newContent : content
+
 const Container = styled.div`
   margin: 60px auto 55px auto;
   padding: 40px 30px 30px 30px;
   width: 502px;
   min-height: 284px;
   background: ${colorGrayscale.white};
-  border-left: solid 1px ${props => props.theme.colors.secondary.support};
+  border-left: solid 1px ${(props) => props.theme.colors.secondary.support};
   ${mq.mobileOnly`
     margin: 40px auto;
     width: 100%;
@@ -35,7 +49,7 @@ const Container = styled.div`
 
 const Title = styled.p`
   display: inline-block;
-  background: ${props => props.theme.colors.secondary.background};
+  background: ${(props) => props.theme.colors.secondary.background};
   padding-right: 2px;
   box-shadow: 5px 15px 0 ${colorGrayscale.white} inset;
   font-size: 22px;
@@ -83,7 +97,7 @@ const Donate = styled.div`
       letter-spacing: 1.3px;
     }
     &:hover {
-      background: ${props => props.theme.colors.secondary.accent};
+      background: ${(props) => props.theme.colors.secondary.accent};
     }
   }
 `
@@ -94,7 +108,14 @@ export default class DonationBox extends PureComponent {
       <Container>
         <Title>{_content.title}</Title>
         {_content.desc.map((p, index) => {
-          return <Text key={`donation-box-desc-${index + 1}`}>{p}</Text>
+          return (
+            <Text
+              key={`donation-box-desc-${index + 1}`}
+              dangerouslySetInnerHTML={{
+                __html: DOMPurify.sanitize(p, { ADD_ATTR: ['target'] }),
+              }}
+            />
+          )
         })}
         <Donate>
           <a